% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turfwaterfall.R
\name{turfwaterfall}
\alias{turfwaterfall}
\title{turfwaterfall}
\usage{
turfwaterfall(catadat)
}
\arguments{
\item{catadat}{From turfR package documentation: "Required. Literal character string representing name of a file in the working directory readable using read.table(data, header=TRUE), or name of a data frame or matrix in R containing TURF data. Rows are individuals (respondents). Columns are (1) respondent identifier, (2) a weight variable, and a minimum of n columns containing only zeroes and ones, each representing an individual item in the TURF algorithm. Respondent identifiers need not be unique and weights need not sum to the total number of rows. In the absence of any weight variable, substitute a column of ones. Ones in the remaining columns indicate that the reach criterion was met for a given item by a given individual. Values other than zero or one in these columns (including NA) trigger an error. data may contain more than n + 2 columns, but any columns in addition to that number will be ignored."

Here is an example of how data should be entered prior to using this package:\tabular{lrrrrrrrrr}{
   ID \tab weight \tab Flavor1 \tab flavor2 \tab flavor3 \tab flavor4 \tab flavor5 \tab flavor6 \tab flavor7 \tab flavor8 \cr
   1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \cr
   2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 0 \tab 0 \tab 0 \tab 1 \tab 1 \cr
   3 \tab 3 \tab 1 \tab 0 \tab 0 \tab 0 \tab 1 \tab 0 \tab 0 \tab 1 \cr
   4 \tab 4 \tab 1 \tab 0 \tab 0 \tab 1 \tab 1 \tab 0 \tab 1 \tab 0 \cr
   5 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \cr
   6 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 0 \tab 1 \tab 1 \cr
   7 \tab 7 \tab 1 \tab 0 \tab 0 \tab 0 \tab 1 \tab 0 \tab 1 \tab 0 \cr
   8 \tab 8 \tab 1 \tab 0 \tab 0 \tab 0 \tab 0 \tab 1 \tab 0 \tab 0 \cr
   9 \tab 9 \tab 1 \tab 0 \tab 1 \tab 1 \tab 1 \tab 1 \tab 0 \tab 1 \cr
   10 \tab 10 \tab 1 \tab 0 \tab 1 \tab 0 \tab 1 \tab 1 \tab 1 \tab 1 \cr
}}
}
\value{
Waterfall plot
}
\description{
turfwaterfall
}
\details{
Plots waterfall graph for reach as you add variables.
}
