% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turfcatatable.R
\name{turfcatatable}
\alias{turfcatatable}
\title{turfcatatable}
\usage{
turfcatatable(catadat)
}
\arguments{
\item{catadat}{CATA Dataframe}
}
\value{
Table of optimal combinations and their reach and frequencies
}
\description{
turfcatatable
}
